+++ packages/contact-center/store/tests/store.ts
@@ -50,6 +50,7 @@ describe('Store', () => {
 
   describe('registerCC', () => {
     it('should initialise store values on successful register', async () => {
+      let date = new Date();
       const mockResponse = {
         teams: [{id: 'team1', name: 'Team 1'}],
         loginVoiceOptions: ['option1', 'option2'],
@@ -57,6 +58,8 @@ describe('Store', () => {
         agentId: 'agent1',
         isAgentLoggedIn: true,
         deviceType: 'BROWSER',
+        lastStateAuxCodeId: 'auxCodeId',
+        lastStateChangeTimestamp: date,
       };
       mockWebex.cc.register.mockResolvedValue(mockResponse);
 
@@ -68,6 +71,8 @@ describe('Store', () => {
       expect(store.agentId).toEqual(mockResponse.agentId);
       expect(store.isAgentLoggedIn).toEqual(mockResponse.isAgentLoggedIn);
       expect(store.deviceType).toEqual(mockResponse.deviceType);
+      expect(store.currentState).toEqual(mockResponse.lastStateAuxCodeId);
+      expect(store.lastStateChangeTimestamp).toEqual(date);
     });
 
     it('should log an error on failed register', async () => {