+++ packages/contact-center/user-state/tests/user-state/index.tsx
@@ -5,14 +5,16 @@ import * as helper from '../../src/helper';
 import '@testing-library/jest-dom';
 
 // Mock the store import
-jest.mock('@webex/cc-store', () => {return {
-  cc: {
-    on: jest.fn(),
-    off: jest.fn(),
-  },
-  idleCodes: [],
-  agentId: 'testAgentId'
-}});
+jest.mock('@webex/cc-store', () => {
+  return {
+    cc: {
+      on: jest.fn(),
+      off: jest.fn(),
+    },
+    idleCodes: [],
+    agentId: 'testAgentId',
+  };
+});
 
 describe('UserState Component', () => {
   let workerMock;
@@ -25,18 +27,26 @@ describe('UserState Component', () => {
     };
     global.Worker = jest.fn(() => workerMock);
     global.URL.createObjectURL = jest.fn(() => 'blob:http://localhost:3000/12345');
+
+    if (typeof window.HTMLElement.prototype.attachInternals !== 'function') {
+      window.HTMLElement.prototype.attachInternals = jest.fn() as any;
+    }
   });
 
   it('renders UserStateComponent with correct props', () => {
     const useUserStateSpy = jest.spyOn(helper, 'useUserState');
-    
-    render(<UserState/>);
+
+    render(<UserState />);
 
     expect(useUserStateSpy).toHaveBeenCalledTimes(1);
-    expect(useUserStateSpy).toHaveBeenCalledWith({cc: {
-      on: expect.any(Function),
-      off: expect.any(Function)
-    }, idleCodes: [], agentId: 'testAgentId'});
+    expect(useUserStateSpy).toHaveBeenCalledWith({
+      cc: {
+        on: expect.any(Function),
+        off: expect.any(Function),
+      },
+      idleCodes: [],
+      agentId: 'testAgentId',
+    });
     const heading = screen.getByTestId('user-state-title');
     expect(heading).toHaveTextContent('Agent State');
   });